# Fixes Type Instability in _evaluate and _derivative