--- a/src/Dierckx.jl
+++ b/src/Dierckx.jl
@@ -257,8 +257,12 @@ function Spline1D(x::AbstractVector, y::AbstractVector,
     return Spline1D(t, c, k, _translate_bc(bc), fp[])
 end
 
-function _evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
-                   x::Vector{Float64}; bc::Int=0)
+# TODO: deprecate this
+_evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int, x; bc::Int=0) =
+    __evaluate(t, c, k, x, bc)
+
+function __evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
+                   x::Vector{Float64}, bc::Int=0)
     bc in (0, 1, 2, 3) || error("bc = $bc not in (0, 1, 2, 3)")
     m = length(x)
     xin = convert(Vector{Float64}, x)
@@ -275,8 +279,8 @@ function _evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
     return y
 end
 
-function _evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
-                   x::Real; bc::Int=0)
+function __evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
+                   x::Real, bc::Int=0)
     bc in (0, 1, 2, 3) || error("bc = $bc not in (0, 1, 2, 3)")
     y = Ref{Float64}(0)
     ier = Ref{Int32}(0)
@@ -293,11 +297,12 @@ end
 
 function evaluate(spline::Spline1D, x::AbstractVector)
     xin = convert(Vector{Float64}, x)
-    _evaluate(spline.t, spline.c, spline.k, xin, bc=spline.bc)
+    __evaluate(spline.t, spline.c, spline.k, xin, spline.bc)
 end
 
 evaluate(spline::Spline1D, x::Real) =
-    _evaluate(spline.t, spline.c, spline.k, x, bc=spline.bc)
+    __evaluate(spline.t, spline.c, spline.k, x, spline.bc)
+
 
 function _derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
                      x::Vector{Float64}; nu::Int=1, bc::Int=0)