--- a/src/Dierckx.jl
+++ b/src/Dierckx.jl
@@ -257,7 +257,7 @@ function Spline1D(x::AbstractVector, y::AbstractVector,
     return Spline1D(t, c, k, _translate_bc(bc), fp[])
 end
 
-# TODO: deprecate this
+# TODO: deprecate this?
 _evaluate(t::Vector{Float64}, c::Vector{Float64}, k::Int, x; bc::Int=0) =
     __evaluate(t, c, k, x, bc)
 
@@ -304,8 +304,14 @@ evaluate(spline::Spline1D, x::Real) =
     __evaluate(spline.t, spline.c, spline.k, x, spline.bc)
 
 
-function _derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
-                     x::Vector{Float64}; nu::Int=1, bc::Int=0)
+# TODO: deprecate this?
+_derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int, x;
+            nu::Int=1, bc::Int=0) =
+    __derivative(t, c, k, x, nu, bc)
+
+
+function __derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
+                     x::Vector{Float64}, nu::Int=1, bc::Int=0)
     (1 <= nu <= k) || error("order of derivative must be positive and less than or equal to spline order")
     m = length(x)
     n = length(t)
@@ -323,8 +329,8 @@ function _derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
     return y
 end
 
-function _derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
-                     x::Real; nu::Int=1, bc::Int=0)
+function __derivative(t::Vector{Float64}, c::Vector{Float64}, k::Int,
+                     x::Real, nu::Int=1, bc::Int=0)
     (1 <= nu <= k) || error("order of derivative must be positive and less than or equal to spline order")
     n = length(t)
     wrk = Vector{Float64}(n)
@@ -345,13 +351,18 @@ end
 #       or should it be integrated with evaluate, above?
 #       (problem with that: derivative doesn't accept bc="nearest")
 # TODO: should `nu` be `d`?
-function derivative(spline::Spline1D, x::AbstractVector; nu::Int=1)
+function derivative(spline::Spline1D, x::AbstractVector, nu::Int=1)
     xin = convert(Vector{Float64}, x)
-    _derivative(spline.t, spline.c, spline.k, xin; nu=nu, bc=spline.bc)
+    __derivative(spline.t, spline.c, spline.k, xin, nu, spline.bc)
 end
 
-derivative(spline::Spline1D, x::Real; nu::Int=1) =
-    _derivative(spline.t, spline.c, spline.k, x; nu=nu, bc=spline.bc)
+derivative(spline::Spline1D, x::Real, nu::Int=1) =
+    __derivative(spline.t, spline.c, spline.k, x, nu, spline.bc)
+
+
+# TODO: deprecate this?
+derivative(spline::Spline1D, x; nu::Int=1) = derivative(spline, x, nu)
+
 
 function _integrate(t::Vector{Float64}, c::Vector{Float64}, k::Int,
                     a::Real, b::Real)