This mini-PR fixes a type instability in `_evaluate`, `_derivative` and `derivative` caused by using kwargs. This can be seen if one makes many calls with just a scalar argument:

```
# OLD TIMINGS 
# evaluate 
#   0.018588 seconds (500.00 k allocations: 15.259 MB, 11.57% gc time)
#   0.017436 seconds (500.00 k allocations: 15.259 MB, 7.48% gc time)
# first derivative
#   0.020274 seconds (600.00 k allocations: 33.569 MB, 6.01% gc time)
#   0.020318 seconds (600.00 k allocations: 33.569 MB, 11.01% gc time)
# second derivative
#   0.027993 seconds (700.00 k allocations: 42.725 MB, 11.26% gc time)
#   0.026323 seconds (700.00 k allocations: 42.725 MB, 10.23% gc time)

# NEW TIMINGS 
# evaluate 
#   0.009857 seconds (200.00 k allocations: 3.052 MB)
#   0.009378 seconds (200.00 k allocations: 3.052 MB)
# first derivative
#   0.012261 seconds (300.00 k allocations: 19.837 MB, 8.79% gc time)
#   0.012860 seconds (300.00 k allocations: 19.837 MB, 6.69% gc time)
# second derivative
#   0.019107 seconds (400.00 k allocations: 28.992 MB, 11.40% gc time)
#   0.018948 seconds (400.00 k allocations: 28.992 MB, 11.45% gc time)
```