@cortner 2017-04-10
I guess this is not a huge performance gain, and would only appear in poorly optimised code, but even this factor 2 matters to me, hence I'd really appreciate if this fix can be accepted.