@kbarbary 2017-04-10
I'm in favor of factor of 2 improvements!